moved AddTaskScreen to components folder
